Memoize computations to prevent rerendering of graph
Whenever you use `map` or any method that produces a copy of the original object, the reference to that object is changed. In the body function of a component, everything is rerun (so `map` is called everytime something in that component changes, giving you a different reference every time). To prevent this - you memoize the computation by providing its direct dependencies - the computation will be the same for the same input (`mentalStates`) so if `mentalStates` doesn't change, the computation will have the same reference, therefore preventing the rerender of the `MentalStatesGraph` component

We also don't need to memo the whole component as the component is only re-rendered when its props change, and the graph is redrawn only when `mentalStateNamesOrdered` or `nextMentalStateNamesOrdered` change